kafka-consumers/requirements.txt
kafka-python
cassandra-driver
requests
